Split test-moment into timestamp and date helpers
