+++ test/integration/admin/sidebar/clinician-sidebar.js
@@ -1,7 +1,7 @@
 import _ from 'underscore';
 
 import { getError } from 'helpers/json-api';
-import { testTs } from 'helpers/test-moment';
+import { testTs } from 'helpers/test-timestamp';
 
 const stateColors = Cypress.env('stateColors');
